Update the "first run" workflow
The big changes here are to bounce the user into their newly-created
subdomain and log them in correctly using User::Relayable and
Sessions::RelaysController.

In the FirstRunsController, I've dropped the FirstRun PORO model and
we're now directly creating the Account and User in the controller
action. Happy to unwind that if folks have a strong opinion.