Script to migrate active storage attachments from Account to Bubble
Note that, after assigning attachments to bubbles, any images not
referenced in a comment will remain with a record type of
"Account". Partly this is because the multi-tenant migration kept all
attachments for all tenants, so some of those attachments belong to
another tenant. But others may have been orphaned by deleted comments.

In any case, cleanup of those orphaned records can be done in a future
migration (maybe when we implement periodic active storage garbage
collection).