Introduce a recursive descent parser for the Fizzy Do command line
Using a hand-written recursive descent parser instead of a LALR parser
generator (like lex/yacc or rex/racc) because IMHO it's easier to
understand, extend, and debug. It also allows us to be a bit more
error-tolerant where we need to.

The AST I'm introducing here is pretty simple, and I'm not tracking
token location because my assumption is that the parsed strings are
all going to be short.

I'll also note that this approach, and some of the API patterns,
borrow from the Ruby "syntax tree" projects, see
https://github.com/ruby-syntax-tree for other living examples.

ref: https://3.basecamp.com/2914079/buckets/37331921/documents/8608517933