beamer-testbed: overhaul
The testbed now comes up in a three-zone configuration. In order to do
this, services must start up in a very particular order, to avoid race
conditions on first boot / initialization:

1. beamer starts up first, and zone configuration happens
2. then the default writer migrations are run, so that the untenanted
   database is in a known zone
3. then the other nodes run migrations,
4. then the web server can be started

Lots of weird things can happen on first boot if things don't happen
in this order. For example ...

- Somewhat obviously, we want to avoid race conditions on creating the
  untenanted+replicated database.
- Less obviously, if the web server comes up before the cable database
  is initialized, it will be an empty database.

We're also running Solid Queue in puma for the moment, and I've made
testbed-specific configuration explicit on the BEAMER_TESTBED env
var (for example, using structured logs).